+++ src/test.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import ReactDOM from 'react-dom/client'
+import { ShadcnTest } from './components/test/ShadcnTest'
+import './index.css'
+
+ReactDOM.createRoot(document.getElementById('root')!).render(
+  <React.StrictMode>
+    <ShadcnTest />
+  </React.StrictMode>,
+)